pyproject.toml
[project]
name = "claude-code-mcp"
version = "2.0.0"
description = "Lean MCP server for Claude Code with voice assistant"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Claude Code MCP Team"}
]

dependencies = [
    # Core server
    "fastmcp>=0.1.0",
    "uvicorn>=0.30.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "aiofiles>=24.0.0",
    
    # GitHub integration
    "PyGithub>=2.0.0",
    
    # Voice client
    "typer>=0.12.0",
    "websockets>=13.0",
    "aiohttp>=3.9.0",
    "pyaudio>=0.2.14",
]

[project.optional-dependencies]
claude = ["claude-agent-sdk"]
dev = ["pytest", "ruff", "mypy"]

[project.scripts]
mcp-server = "src.main:main"
voice = "client.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
strict = false
